Reorganized files, grouped related components
