+++ src/components/Worker/ViewNResult.tsx
